+++ lib/tapioca/internal.rb
@@ -18,7 +18,7 @@ require "tapioca/executor"
 require "tapioca/symbol_table_parser"
 require "tapioca/symbol_loader"
 require "tapioca/gem/events"
-require "tapioca/compilers/gem/listeners"
+require "tapioca/gem/listeners"
 require "tapioca/compilers/requires_compiler"
 require "tapioca/compilers/symbol_table_compiler"
 require "tapioca/dsl/pipeline"
